Solve Day 23 :star:
The original solution for part 1 was built using byte arrays. The fixed
size nature of them prevented from code reuse for part 2, where room
size changed.

Some problem terms allowed to use a single integer for both parts. The
speed was not affected greatly, and there are still some places for
improvement. That would complicate the solution of course.